# Optimize internal clip_values function